Overhaul to Find a Space cards (#599)
* Shimmy & shake for spaces cards #591

Use link card and utilize multiple anchors to avoid nested anchors. Not
completely thrilled with this approach because it leads to certain areas of
whitespace that appear clickable but are misleading. Might try adding click
event to .description div to address this.

* Bump All Spaces to the front of the class

First in subnav list as suggested during staff show 'n tell.

* New icon for quiet study spaces

Utilzing Font Awesome's stackable icons. Don't know why I didn't think of this
earlier.

* A cleaner JS version of shimmy & shake #591

As eluded to in 2530d95. This relies on a click event for the card's .content
div and adding a listener that stops event bubbling on any nested anchors
within.

* Comment out hardcoded hours for spaces #590