Use url & file loaders for Font Awesome fonts #492
Drop the nonsense of copy-webpack-plugin, as called out in 5f57190.

Now semantic-ui's icon component (port of Font Awesome) works as expected which
means we have the delete option on the multiple value dropdown component and the
previous/next navigation on the datepicker/calendar component (#491), to name a
few.

Added bonus is that this should also address the second false positive from the
recent AppSpider scan (#465), Nginx NULL code vulnerability, related to serving
fontawesome-webfont.ttf.