Use Font Awesome CDN (#621)
My previous attempt to refactor the Font Awesome font handling in 3c4bbc0 was
too good to be true. Browser cache was spoofing expected functionality when
reality was far from it. I could go back to using copy-webpack-plugin, but that
would put us in the same spot with AppSpider scan #465.

So...keep it simple and go with the CDN until further notice. Should keep
everyone happy in the meantime.