Apply stickyheaders after softwareList is created
This is the key to the bug in #606 and the reason why I was not able to reliably
reproduce this is because when hosting the site on wagon, `$(window).load` would
trigger after the Vue component was updated, but in the opposite order when
hosting on engine. We can ensure the proper order by utilizing Vue's watch
method.