Editable file fix for section photo #466 (#707)
Went over this again, playing close attention to this mini-guide in the docs:
https://locomotive-v3.readme.io/docs/tags#section-how-to-use-editable-image-with-a-default-value-

Still having issues when trying to retreive the editable file via page.children
(on landing -- still nil) but seem to be in good shape for section layout. Let's
see what happens when we deploy to dev Engine (local Engine was good for
section, local wagon was good for section AND landing).