Academic courses toggle for events calendar #720
Default is to NOT INCLUDE academic courses in the events calendar.

Non-functional but consider it a prototype to indicate the direction we want to
head in when re-introducing R25 data back into the mix.

Plus some overall cleanup of events calendar layout for all breakpoints around
the top of the page.