Hotfix - filter reservable spaces for Engine #608
Ideally, we would use the `reservation_system` field to filter but Engine chokes
because it's expecting a BSON objectid. I couldn't figure out an operator to use
on a relationship field that could simply test for existence of relationship, so
falling back to test for the existence of the `reserve_sys_id` field via the
`gt` operator since it's an integer field.

Awkward, but works no matter whether the site is served by Wagon or Engine.

List of operators for `with_scope` tag:
https://github.com/locomotivecms/steam/blob/master/lib/locomotive/steam/liquid/tags/with_scope.rb#L19