Revert #757: put megamenu touch JS in its place
This should have been a non-issue, but both Events apps (homepage & calendar)
are currently not using babel-loader and appear to be requiring a second copy of
jQuery. The result is that the touch handling of the megamenu is broken on the
homepage and the calendar because the `doubletaptogo` plugin is not resolvable
as a function.

Need to get to the bottom of this, but hit a dead end and have a couple
of final features that need to be implemented for the launch.

This reverts commit 7b6d3549dd8b9342faa3628d2c0908f641004c1f.