Revert megamenu touch refactor (#761)
* Remove defunct jquery-local.js

This is a holdout from the original approach to bundling jQuery via Webpack
(7576b2f). Should have been dropped as part of 62f301a.

* Revert #757: put megamenu touch JS in its place

This should have been a non-issue, but both Events apps (homepage & calendar)
are currently not using babel-loader and appear to be requiring a second copy of
jQuery. The result is that the touch handling of the megamenu is broken on the
homepage and the calendar because the `doubletaptogo` plugin is not resolvable
as a function.

Need to get to the bottom of this, but hit a dead end and have a couple
of final features that need to be implemented for the launch.

This reverts commit 7b6d3549dd8b9342faa3628d2c0908f641004c1f.

* Ensure megamenu's z-index trumps all others

Carousel component's nav is also using `$above-content` variable, which means
the megamenu did not cover it -- a bug on smaller screens.