Libcal widget works in both views of staff dir
Turns out this never worked properly on the staff directory due to having the
consultation button for each staff member rendered two times due to the two
views (grid vs table) -- the table is hidden via JS on page load, but the
duplication of IDs caused only the first button in the DOM (the table) to work.

Ensuring unique IDs for the buttons addressed this problem.