Temp punt on desktop availability via Labstats
Labstats API went down and/or was unresponsive and/or there were issues with
certain nameservers (dnsexit.com?) properly resolving online.labstats.com. See
Alan for more details.

This is a quick fix and we should probably update to latest Engine & Steam RCs
so we can reap the benefit of the recent enhancment to the `consume` tag:

locomotivecms/steam@e0b85a3