Still no go for section photos on landing pages :(
So the originally reported bug from locomotivecms/engine#1172 was
addressed in Steam `1.3.0.rc1` (locomotivecms/steam@96d3b79), which has
been an Engine dependency since `3.3.0.rc1`.

Unfortunately, while the value of an `editable_element` retrieved from a
page drop is no longer `nil`, what's returned for an editable file is
simply a string of the filename. This results in the exact same behavior
as prior to this fix where the file only renders when the site is hosted
by Wagon.

It would seem that for an editable file, the preferred content returned
within the drop would be the `element.base_url` to ensure that Steam can
render the file regardless of whether the site is hosted by Wagon or
Engine. I've added a comment to locomotivecms/engine#1172, asking Didier
if there's anyway to currently achieve this or if it will require a new
filter (and underlying service -- `editable_file_url_service`?).

Stay tuned... #466, #707