Move Webpack bundle optimization to prod config
Along the way, drop Webpack plugins introduced in #839:

* optimize-css-assets-webpack-plugin
* uglifyjs-webpack-plugin

Turns out, Webpack includes UglifyJSWebpackPlugin via
`webpack.optimize.UglifyJsPlugin` [1]. Yes, even Webpack 1 [2]. Not only
that, but UglifyJS is able to minimize CSS via cssnano when using the
`css-loader` [3], which we are =)

Tested and confirmed that the minified CSS was equivalent when using
included UglifyJS vs the combo of the two plugins. As a matter of fact,
`optimize-css-assets-webpack-plugin` was redundant from the get-go.

Still need to add script to package.json to build production assets.

[1] https://webpack.js.org/plugins/uglifyjs-webpack-plugin
[2] https://webpack.github.io/docs/list-of-plugins.html#uglifyjsplugin
[3] https://github.com/webpack/docs/wiki/optimization#minimize