Bugfix: Wrap unlinked spotlight img in <span> #902
This bug was original introduced when switching to lazy-load way back in
1eb881b as part of #846. Turns out Slick's `data-lazy` attribute
requires the image to be wrapped so it's not a direct child of the
element that's targeted for the carousel. This explains why linked
spotlights were rendering properly (wrapped within an anchor <a>).

Also drops non-existent `image_url` liquid filter which appears to have
been hanging around since spotlights were first implemented.

P.S. This addresses the dupe cul-it/mann-sitefeedback#53.