Update equipment (#924)
* Add reserve items to equipment page

* Made grid responsive

* Add accesibility link

* Add temporary hack to shame css component

* Remove underline links on buttons

* Create equipment type array & loop through

Went with the singular variant for all.

* Use children from pages drop to build equip list

* Dynamically build reserve URLs via Actions API

Introduces two new fields to sensitive data site-metafields due to
authentication requirements for LibCal v1.1 API.

Includes debugging info we should be removed before we merge.

Todo:

* update equipment pages (title, slug & handle should be singular)
  -- discussed in Slack
  -- LibCal categories have already been updated
  -- none of the reserve URLs will work properly until this happens
  -- okay to make changes to pages locally, but don't commit

* remove debugging info

* add logic to handle categories/types that are not reservable

* Build equipment list & links dynamically via children pages & LibCal API (#921)

* Create equipment type array & loop through

Went with the singular variant for all.

* Use children from pages drop to build equip list

* Dynamically build reserve URLs via Actions API

Introduces two new fields to sensitive data site-metafields due to
authentication requirements for LibCal v1.1 API.

Includes debugging info we should be removed before we merge.

Todo:

* update equipment pages (title, slug & handle should be singular)
  -- discussed in Slack
  -- LibCal categories have already been updated
  -- none of the reserve URLs will work properly until this happens
  -- okay to make changes to pages locally, but don't commit

* remove debugging info

* add logic to handle categories/types that are not reservable

* Sync equip pages name changes (singular) from prod

We still have to add a page for the thermal imager.

* Sync camera page title change from prod

Now `DSLR Camera`.

* Refactor Libservices equipment type logic (#923)

* Refactor Libservices equipment type logic

Cleanup on aisle 6. Still a mess and still hardwired, but should hold us
over for the interim solution proposed in #886.

* Drop all garbage equipment types while we're at it

Cleanup on aisle 7? More Libservices tomfoolery...

* Remove LibCal data debug
Add logic for disabling reserve buttons when equipment is not reservable

* Use capture liquid tag for reserving equipment

* GPS model is Garmin eTrex 30x

* Drop old loanable equipment list setup snippet

No longer used.

* Quick tooltip to explain disabled reserve button

Not all equipment is reservable. Would have liked to use semantic-ui css
tooltip but that got tricky with opacity and z-index (side-effect of
disabled attribute).

* Cleanup policies and reserve helper callout

Simpler and sweeter. Also link to hours.

* Fix some sass lint warnings

* Drop temporary style hiding page content

Definitely don't want this in production even though for the time being
we plan on managing most of the copy within the snippet. But this will
be refactored to use editable-elements if this interim solution sticks
around for a bit.

* Expand scope of style for underlining links

Target the `.page-content` and `.page-content--dynamic` divs instead of
simply paragraphs. This provides better coverage due to dirty/messy
markup introduced by copy/pasting into back-office wysiwyg and also has
added benefit of including linked headings.

Also move the rule to `_a11y` component, where it should have been from
the beginning when introduced in 117427c as part of #907.