Future proof the monthly hours view
Instead of maintaining a list of LibCal sub locations (aka departments)
that should not be rendered as originally introduced in #472 [1], use
the power of Sass and the ampersand to target the only sub location
(lobby) we care about as the exception to the rule.

The second and final piece to #939.

[1] https://github.com/cul-it/mann-wagon/pull/472/commits/c8d727aa336dfd9f1ed920e6241b1943fa7bfd19#diff-81dd4da25cc3000181a11037926584e0