Fall back to copied semantic-ui-calendar src
A friendly reminder that this is the quick-n-dirty approach. Remove
semantic-ui-calendar dependency and copy v0.0.3 dist files [1] directly
into app.

This will be the baseline which will then be overwritten by brute force
to replace <i> with <span> for all embededded instances of the icon
component.

This essentially reverts a good chunk of 52ec386 (within #804), where we
added semantic-ui-calendar dependency via yarn in lieu of manually
downloaded version.

[1] https://github.com/mdehoog/Semantic-UI-Calendar/releases/tag/0.0.3