Events calendar gracefully handles missing data (#1009)
* Events calendar fails gracefully w/ missing data

Traced the underlying issue down to a previous workaround introduced
in #774 as part of the original site launch back in Oct 2016 which
attempted to catch null/empty values for several fields and reset the
events array. This came to a boil this week when the events calendar was
stuck in an endless loop, both on the home page and full events calendar
page, due to missing/empty values for advertising with CU Events
calendar, event name and event description

This commit introduces a failsafe to ensure the events calendar degrades
gracefully. The patron submitted values embedded in the description field
of the LibServices relay feed are parsed and missing values are handled
explicitly. To be clear, all of these fields are required on the public
form, but staff with LibCal accounts often bypass the form and add a
reservation directly with minimal info (date and time) :(

We first encountered this missing values bug in Jan 2018 during the
LibCal E&S migration with event_type field. I should have listened to
myself at that point and implemented this failsafe for the rest of the
.match() occurrences at that time #938. It's actually kind of incredible
that we haven't run into this endless loop scenario prior to this point.

* Reorganize code for sanity

Rename variable introduced in previous work to have sensible defaults
for event_type (#938) for consistency with the rest of the failsafe
handling/parsing from this PR.

* Apply the same logic & handling to homepage

Unfortunately this is almost a verbatim dupe between the main Vue app
for events calendar page and the embedded Vue app on the homepage.

So. Not. DRY.