Leverage categories for LibCal API space endpoint (#1014)
* All indiv, small & large group rooms in LibCal

Drop any logic dealing with LibServices for these room types now that
they are all reservable. Down with loans!

* Single http request for LibCal space availability

Limit to single http request to retrieve availability of spaces, even if
they are actually comprised of multiple rooms in LibCal.

Addresses #1008.

Most drastic example is individual study rooms (12 rooms total). This
PR reduces the number of http requests from up to 12 (depending on
availability) to 1, regardless of availability.

If individual room IDs are specified:

* Create array
* Loop through array, making http request per room

If individual room IDs are NOT specified

* Use the reserve_sys_id as category
* Make single http request and retrieve availability for all rooms
  within category

It should also be noted that this commit fixes the next availability bug
where the logic for determining the next available space fell apart when
dealing with more than two rooms (whether via a category or via listed
space IDs).