Teams ftw (#1016)
* New content type for teams

The successor to liaisons and all things good and wholesome.

* Dynamic content for disciplinary research support

DReSS full team photo and description up top, followed by each sub-team
name with description and head shots of members that link back to staff
profile.

* Remove divisions from staff directory

But identify team members.

* Render team membership on home page

Within Meet Our Experts section.

* Refactor Meet our Experts for teams

Remove department/liaison filtering and replace with ability to filter
based on team membership.

* Prototype of new callouts for home page

As requested by Erica.

* Library Instruction – Submit your request

* Our collection – Suggest a recommendation

* Research Support – Consult with our experts

* Temp vector images for home callouts

As a separate commit so it can be easily reverted once we come to a
final decision.

* Preferred pronouns for staff

* Refactor DReSS page based on Erica & Tobi feedback

* drop DReSS full team photo and description

* add subteam photos

* drop subteam description

* link to filtered view of Meet Our Experts page for subteam links

* add "Make an Appt." button – unable to prepopulate Libwizard form :(

This also refactors the Meet Our Experts page to respond to requests
that include team filters. In the process of tweaking the behavior of
the initial request, I simplified the "View all experts" link from the
home page.

* Introduce team-link snippet to DRY up team links

Consistent handling of linking to a team filtered view of the Meet Our
Experts page.

* Research Support goes to the head of the class

First in the homepage callouts as per Erica's request.

* Manual sort for teams from backoffice

To allow for Erica's desired order of DReSS sub-teams.

* Request btns (instruct & consult) for meet experts

As per Erica's request.

* Link to filtered catalog search for Our Collection

Along with "Explore our resources" call to action.

As per Tobi's request.

* Book a Consultation: back with a vengeance

As per Tobi's demand ;)

* Daisy's illustrations for homepage callouts

Nothing beats in-house talent =)

* Consultation callout should link to DReSS page

As per Erica's request.