Ensure LibCal events array is marked as updated
Even if there are no upcoming events, the empty result set needs to be
stored to the libcalEvents array since it is being watched [1] to
indicate that the fetch is complete.

This step was inside a test for array length and normally wouldn't
present as an issue since it is unlikely we would have no LibCal
reservations in the next 35 days, but these are unusual times.

It should be noted that this resulted in a broken state of the Events
calendar (forever loading) which was mistakenly attributed to the May
2020 ADDTrust External CA Root Expiration [2], but this is unrelated and
simply unfortunate timing.

[1]: https://v1.vuejs.org/api/#watch
[2]: https://calnetweb.berkeley.edu/calnet-technologists/incommon-sectigo-certificate-service/addtrust-external-root-expiration-may-2020#affected