Fall 2020 course reserves (#1076)
* Update course reserves to show text if due date isn't null and isn't available, hide overdue if item is not checked out

* Add "Online" qualifier to due date to not display it double

* Revert recent tweaks to course reserves logic

These are only necessary because of some hardcoding introduced to the
logic in April 2020 via #1022. It's preferable to roll back those
earlier changes as well to limit maintenance requirements moving
forward.

More details to come in follow up commits here...

> This reverts commits ecf9ac3 & 0dbf41c which were originally
cherry-picked from #1070 (33e70d6 & 5e9b86d).

* Revert April 2020 course reserves logic tweaks

Backs out of the hardcoded logic for temporary unavailability of course
reserve items due to library closure during the COVID-19 pandemic.

This reverts kevinkidwell/mann-wagon@e252800 which was merged as part
of #1022. The one caveat is that the switch from Blackboard to Canvas
for online items is retained =)

* New course reserves logic for Fall 2020 (COVID-19)

* Always display dueDate string if item is not online
* Only display overdue message if item is checked out

These two changes allow for the custom COVID-19 status messages set at
the course reserves API to be displayed without interference.

The nested v-else and v-if is necessary since v-else-if was introduced
in Vue 2.1.0 [1].

[1]: https://vuejs.org/v2/guide/conditional.html#v-else-if

Co-authored-by: Kevin Kidwell <kekidwell@gmail.com>