Bugfix: Ensure Localist events are rendered (#1079)
* Test room_number for truthy

Instead of negative test for empty string, which misses NULL.

This was preventing the events calendar from displaying any/all events.

Based on the fact that all events are virtual, it's not surprising that
the room_number field in Localist is empty. There's a good chance this
gave us problems in the past but we missed it due to the volume of
events. Now that the event count is way down and Localist is the only
data source (LibCal space reservations are currently inactive), there
was nowhere for this bug to hide =)

* Test for truthy instead of negative empty string

Throughout events app, follow suit and reasoning from previous commit
c347d4f.

* Fix links to site feedback form